Added addContextInitializationParameter method. Adds context parameters that are accessible within the application with servletContext.getInitParameter
Deprectated addContextEnvironmentAndResourceFromFile method and
replaced it with setContextFile. The class no longer tries to read this
file it simply passes the location of the file to Tomcat and he will
then read the file during startup. As a result every configuration that
is allowed in the context xml file is now supported.